% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqi_equation.R
\name{aqi_equation}
\alias{aqi_equation}
\title{Calculate Air Quality Index (AQI) and classification}
\usage{
aqi_equation(pollutant, concentration)
}
\arguments{
\item{pollutant}{Either "pm2.5" or "ozone"}

\item{concentration}{A vector of concentrations of that pollutant,
in ug/m^3 for pm2.5 or in ppb for ozone}
}
\value{
A data frame with the columns Concentration, AQI, Class, and Color
}
\description{
takes in a vector of concentrations (of either PM2.5 or ozone)
and calculates the AQI, classification (ranging from "Good" to "Hazardous"),
and associated color (ranging from Green to Maroon)
}
\examples{
aqi_equation("pm2.5", seq(0,300,10))
aqi_equation("ozone", seq(0,0.2,0.0075))
}
\author{
Ellen Considine
}
