% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_requests.R
\name{data_requests_to_markdown}
\alias{data_requests_to_markdown}
\title{Convert a downloaded Qualtrics csv to a project list markdown file}
\usage{
data_requests_to_markdown(file = "NSAPH Health Data Request Form.csv",
  outfile = "project_list.md")
}
\arguments{
\item{file}{path to the csv file containing data requests}

\item{outfile}{name of the file to be produced}
}
\description{
Convert a downloaded Qualtrics csv to a project list markdown file
}
