% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_requests.R
\name{download_data_requests}
\alias{download_data_requests}
\title{Download data request csv from qualtrics}
\usage{
download_data_requests(out_path = ".")
}
\arguments{
\item{out_path}{path to save data in}
}
\description{
Download data request csv from qualtrics
}
\details{
Your qualtrics API access token must be exported as QUALTRICSTOKEN to access the survey results.
Also your quatrics user id needs to be exported as QUALTRICSUSERID
}
