% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor_tools.R
\name{gen_submit}
\alias{gen_submit}
\title{Generate submit file template in a directory}
\usage{
gen_submit(path = ".", out_name = "r.submit")
}
\arguments{
\item{path}{path to the directory where the submit file should be placed}

\item{out_name}{file name for the submit file, should end in ".submit"}
}
\description{
Generate submit file template in a directory
}
