% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{int_to_zip_str}
\alias{int_to_zip_str}
\title{Convert zips stored as ints to 5 digit strings}
\usage{
int_to_zip_str(zip)
}
\arguments{
\item{zip}{zipcode represented as an integer}
}
\description{
Convert zips stored as ints to 5 digit strings
}
\details{
This function initially assumes that the input is intended to be  a 5 digit zipcode.
    If the initial conversion is more than 5 characters, the input is assumed to be intended to
    be a 9 digit zipcode. The function still returns the 5 digit equivalent of the 9 digit
    zip code.
}
