% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_docs.R
\name{list_available_data}
\alias{list_available_data}
\title{Get list of available variables}
\usage{
list_available_data(src, filter.vars = NULL)
}
\arguments{
\item{src}{the type of data to list}

\item{filter.vars}{named list of variables and values to prefilter the data list}
}
\description{
Get list of available variables
}
