% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rce_users.R
\name{list_users}
\alias{list_users}
\title{Return a vector of users in a group}
\usage{
list_users(group = "ci3_confounders")
}
\arguments{
\item{group}{group to list the members of}
}
\description{
Return a vector of users in a group
}
