% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{medicaid_csv_dir_to_fst}
\alias{medicaid_csv_dir_to_fst}
\title{Convert a directory of CMS Medicaid MAX data .csv files to R's fst storage format.}
\usage{
medicaid_csv_dir_to_fst(path, dict, outdir, pattern)
}
\arguments{
\item{path}{The path to the Medicaid data directory for a single year.}

\item{dict}{A dictionary as produced by \code{\link{parse_fts}}.}

\item{outdir}{The directory to write the `.fst` files to.}

\item{pattern}{Regular expression to match `.csv` files to convert.
Useful for separating `ps` (patient summary) from `ip` (inpatient) data,
as these are stored in the same directory in the origina CMS data.}
}
\description{
CMS ships Medicaid data as a set of `.csv`` files, one for each state/year. These `.csv``
files are distributed along with `.fts` (File Transport Summary) files containing file
and variable level metadata. This function converts these to `.fst` format to allow
faster selective IO from R. It does not write over any existing `.fst` files, so
if you want to replace existing `.fst` files you must delete them first.
}
