% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{parse_fts}
\alias{parse_fts}
\title{Parse a .fts file to produce a data dictionaryh.}
\usage{
parse_fts(path)
}
\arguments{
\item{path}{The path to the .fts file.}
}
\description{
Parse a .fts file to produce a data dictionaryh.
}
