% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condor_tools.R
\name{prep_batch}
\alias{prep_batch}
\title{Create a default batch file infrastructure in a directory}
\usage{
prep_batch(path = ".")
}
\arguments{
\item{path}{path to the directory where the structure should be created}
}
\description{
Create a default batch file infrastructure in a directory
}
