% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read Shareded Data}
\usage{
read_data(path, years, columns, nthreads = get_cpus())
}
\arguments{
\item{path}{path to the directory containing the sharded data}

\item{years}{vector of years of data to read}

\item{columns}{vector of strings listing the columns of data to select}

\item{nthreads}{number of threads to use in processes. Default detects requested cpus on rce to force good behavior

Prototype assumes fst data and that data is broken out by year, year used as suffix}
}
\description{
Read Shareded Data
}
