% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{reverse_string}
\alias{reverse_string}
\title{Reverse a vector of strings}
\usage{
reverse_string(str)
}
\arguments{
\item{str}{A vector of strings to be reversed}
}
\value{
a character vector
}
\description{
Reverse a vector of strings
}
