% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{set_threads}
\alias{set_threads}
\title{Set Data Table and FST thread count}
\usage{
set_threads(nthreads = get_cpus())
}
\arguments{
\item{nthreads}{the number of threads to use}
}
\description{
Set Data Table and FST thread count
}
