% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_function.R
\name{template_function}
\alias{template_function}
\title{Sum Three Numerical Values}
\usage{
template_function(a, b, c)
}
\arguments{
\item{a}{first numerical parameter}

\item{b}{second numerical parameter}

\item{c}{third numerical parameter}
}
\value{
The sum of \code{a}, \code{b}, and \code{c}
}
\description{
This functions sums three numerical values.
}
\examples{

s1 <- template_function()

}
\author{
Naeem Khoshnevis, \email{nkhoshnevis@g.harvard.edu}
}
