% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{int_to_zip_str}
\alias{int_to_zip_str}
\title{Convert zips stored as ints to 5 digit strings}
\usage{
int_to_zip_str(zip)
}
\arguments{
\item{zip}{zipcode represented as an integer}
}
\description{
Convert zips stored as ints to 5 digit strings
}
