% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_manipulation.R
\name{refactor_files}
\alias{refactor_files}
\title{refactor multiple files at once}
\usage{
refactor_files(path, orig, replace, pattern = NULL)
}
\arguments{
\item{path}{directory containing files to be refactored}

\item{orig}{original text to be refactored}

\item{replace}{text to replace the refactored phrase with}

\item{pattern}{pattern for further identifying files to be refactored}
}
\description{
refactor multiple files at once
}
